rewrite

+++ api/service.py
@@ -22,18 +22,18 @@ def Send_email(user_email, message):
     )
 
 
+
+
 def create_or_delete(classmodel, **kwargs):
-    try:
-        obj = classmodel.objects.get(**kwargs)
-        obj.delete()
-        return Response({'result': f'обьект {classmodel.__name__} удален'}, status=status.HTTP_200_OK)
+    obj, created = classmodel.objects.get_or_create(**kwargs)
 
-    except:
-        obj = classmodel.objects.create(**kwargs)
+    if created:
         obj.save()
-        return Response({'result': f'обьект {classmodel.__name__} создан'}, status=status.HTTP_200_OK)
-
-
+        return Response({'result': f'объект {classmodel.__name__} создан'}, status=status.HTTP_200_OK)
+    else:
+        obj.delete()
+        return Response({'result': f'объект {classmodel.__name__} удален'}, status=status.HTTP_200_OK)
+    
 def generate_email_code():
     code = random.randrange(start=10000000, stop=99999999)
     return code
